Prevent Update Cycle On Unsortable Header Click
Render header cells without an onClick handler, if they are unsortable.
This prevents an useless update cycle, caused by clicking the header
cell of an usortable column.
To make this possible the signature of the thead field in the
customization record is changed to accept a list of attributes instead
of a single attribute, which enables to pass an empty list if no onClick
handler should be added